requirements.txt
APScheduler==3.10.4
asgiref==3.8.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==5.0.6
django-compat==1.0.15
django-rest-enumfield==0.2.0
djangorestframework==3.15.1
drf-spectacular==0.27.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
numpy==1.26.4
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pluggy==1.5.0
pytest==8.2.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.2
rpds-py==0.18.1
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
